pyproject.toml
[project]
name = "neuralstyletransfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
